samples/rag-storage-document-loaders/requirements.txt
langchain-azure-storage==1.0.0
langchain-azure-ai==1.0.3
langchain-community==0.4.1
pypdf==6.2.0
python-dotenv[cli]==1.1.1